feat: move file to appointment service folder
