+++ src/services/larksuite/appointment/appointment-service.ts
@@ -12,7 +12,7 @@ import FrappeClient from "src/frappe/frappe-client";
 import LarksuiteService from "src/services/larksuite/lark";
 import { fetchLeadInfoByPhoneNumber } from "frappe/lead";
 import { mapLarkToFrappe } from "frappe/utils/utils-lark";
-import { saveAppointmentToPrismaDb } from "src/prisma/utils/appointment-save";
+import { saveAppointmentToPrismaDb } from "services/larksuite/appointment/appointment-save";
 import { getDocumentAttachments, removeFileAttachment } from "frappe/attachment";
 export default class AppointmentService {
   env: any;
